requirements.txt
langchain
langgraph
pinecone
openai
python-dotenv
langchain-community
langchain-chroma
langchain-openai
yahooquery
snowflake-connector-python
matplotlib
tavily-python
yfinance
sqlalchemy
snowflake-sqlalchemy
boto3